+++ cs_api_io/cs_api_io.mli
@@ -1,6 +1,7 @@
 val send_request :
-  ?verify:bool
+     ?verify:bool
   -> Api.Request.t
-  -> (((Cohttp.Response.t * Cohttp_lwt.Body.t), string) result) Lwt.t
+  -> (Cohttp.Response.t * Cohttp_lwt.Body.t, string) result Lwt.t
 
-val get_response : (Cohttp.Response.t * Cohttp_lwt.Body.t) -> ((string, string) result) Lwt.t
+val get_response :
+  Cohttp.Response.t * Cohttp_lwt.Body.t -> (string, string) result Lwt.t